feat: move workspace switcher into composer toolbar
